requirements.txt
Django==1.11.5
MySQL-python==1.2.5
PyJWT==1.5.3
amqp==2.2.2
argparse==1.2.1
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
daemon==1.1
django-unixdatetimefield==0.1.6
idna==2.6
kombu==4.1.0
oauthlib==2.0.4
python-openid==2.2.5
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
vine==1.1.4
wsgiref==0.1.2
zeroc-ice==3.7.0
